Incomplete Version of my stuff
Added in the ball movement related stuff, there are some bugs here and
there, but I figured you wanted something earlier in the day while I
worked on it throughout friday. You need to obtain a certain speed in
order to do a complete roll to the top and then roll down successfully.